refactor: runtime facade interface
